* transforms unknown dynamic invoke into NoSuchMethodError currently only StringConcat and Lambda bootstraps are recognized and cause DynamicInvoke to be transformed/dessugared.
in other cases DynamicInvoke instruction will stay in place and will cause compilation type exception:
> Java.lang.ClassCastException: soot.jimple.internal.JDynamicInvokeExpr cannot be cast to soot.jimple.InstanceInvokeExpr

Issue was risen in gitter channel in scope Scala/desirialize labda being inserted in all classes. All these classes were failed to compile. Even if this functionality is not used (https://github.com/scala/scala/pull/4501).

As a workaround changed how InvokeDynamic is being handled:
- introduced single InbokeDynamicCompilerPlugin;
- LabdaPlugin and StringconcatRewriter plugins are made as delegates of InbokeDynamicCompilerPlugin;
- all not recognized InvokeDynamic are now translated into NoSuchMethodError exceptions
